+++ src/Nethermind/Nethermind.Taiko/Rpc/TaikoEngineRpcModule.cs
@@ -272,7 +272,7 @@ public class TaikoEngineRpcModule(IAsyncHandler<byte[], ExecutionPayload?> getPa
                     }
 
                     // For Surge, filter out any transaction with very high gas limit
-                    if (surgeConfig.MaxGasLimitRatio > 0 && tx.GasLimit > tx.SpentGas * (ulong)surgeConfig.MaxGasLimitRatio)
+                    if (surgeConfig.MaxGasLimitRatio > 0 && tx.GasLimit > tx.SpentGas * surgeConfig.MaxGasLimitRatio)
                     {
                         Restore(snapshot, gasUsedBefore);
                         while (i < txSource.Length && txSource[i].SenderAddress == tx.SenderAddress) i++;
